New Proposals working
A non-DAO member can submit a new proposal to the DAo that a member of
the DAO can approve and put to a vote. I left the request token amount
in the backend code as we'll probably want to implement that to issue an
NFT in the future.

TODO: make a form that the user can fill out that is parsed to json that
the registry can use later to display info about the proposal if it
passes.

Whenever I change TokenRequest.sol to RegistryProposal the template
fails to deploy. Will have to investigate more in that area.